🚚 [projects] Move type alias `CreateProject` to `common` module
